requirements.txt
# Core dependencies for computer vision image classification
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
tensorflow>=2.8.0
fastapi>=0.78.0
uvicorn[standard]>=0.18.0
pydantic>=1.9.0
python-multipart>=0.0.5

# Data processing and analysis
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
bandit>=1.7.0

# Optional: PyTorch alternative
# torch>=1.11.0
# torchvision>=0.12.0

# Logging and configuration
python-dotenv>=0.19.0
structlog>=22.1.0

# API documentation
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4